CIF-1473: Refactor product options component
