+++ react-components/src/components/BundleProductOptions/__test__/multiSelect.test.js
