+++ react-components/src/components/BundleProductOptions/bundleProductOptions.js
@@ -12,7 +12,6 @@
  *
  ******************************************************************************/
 import React, { useState } from 'react';
-import ReactDOM from 'react-dom';
 import { instanceOf } from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import Checkbox from './checkbox';
@@ -171,7 +170,7 @@ const BundleProductOptions = props => {
 
     if (optionsContainer !== null) {
         const { sku } = optionsContainer.dataset;
-        return ReactDOM.createPortal(
+        return (
             <>
                 {bundleState === null ? (
                     <section className={'productFullDetail__section ' + classes.customize}>
@@ -180,78 +179,77 @@ const BundleProductOptions = props => {
                         </Button>
                     </section>
                 ) : (
-                    <>
-                        {bundleState.options.items.map(e => (
-                            <section key={`item-${e.option_id}`} className={'productFullDetail__section '}>
-                                <h3 className="option__title">
-                                    <span>{e.title}</span>{' '}
-                                    {e.required && <span className={classes.required_info}> *</span>}
-                                </h3>
-                                <div>{renderItemOptions(e)}</div>
-                            </section>
-                        ))}
-                        <section className="productFullDetail__section">
-                            <h3>
-                                <span className={classes.required_info}>
-                                    * {t('product:required-fields', 'Required fields')}
-                                </span>
-                                <span className={classes.customization__info}>
-                                    {t('product:customization-price', 'Your customization')}: {getTotalPrice()}
-                                </span>
-                            </h3>
-                        </section>
-                        <section className="productFullDetail__quantity productFullDetail__section">
-                            <h2 className="productFullDetail__quantityTitle option__title">
-                                <span>{t('cart:quantity', 'Quantity')}</span>
-                            </h2>
-                            <div className="quantity__root">
-                                <span className={'fieldIcons__root ' + classes.select_icons}>
-                                    <span className="fieldIcons__input">
-                                        <select
-                                            aria-label="product's quantity"
-                                            className="select__input field__input"
-                                            name="quantity">
-                                            <option value="1">1</option>
-                                            <option value="2">2</option>
-                                            <option value="3">3</option>
-                                            <option value="4">4</option>
-                                        </select>
+                        <>
+                            {bundleState.options.items.map(e => (
+                                <section key={`item-${e.option_id}`} className={'productFullDetail__section '}>
+                                    <h3 className="option__title">
+                                        <span>{e.title}</span>{' '}
+                                        {e.required && <span className={classes.required_info}> *</span>}
+                                    </h3>
+                                    <div>{renderItemOptions(e)}</div>
+                                </section>
+                            ))}
+                            <section className="productFullDetail__section">
+                                <h3>
+                                    <span className={classes.required_info}>
+                                        * {t('product:required-fields', 'Required fields')}
                                     </span>
-                                    <span className="fieldIcons__before"></span>
-                                    <span className="fieldIcons__after">
-                                        <span className="icon__root">
-                                            <svg
-                                                xmlns="http://www.w3.org/2000/svg"
-                                                width="18"
-                                                height="18"
-                                                viewBox="0 0 24 24"
-                                                fill="none"
-                                                stroke="currentColor"
-                                                strokeWidth="2"
-                                                strokeLinecap="round"
-                                                strokeLinejoin="round">
-                                                <polyline points="6 9 12 15 18 9"></polyline>
-                                            </svg>
+                                    <span className={classes.customization__info}>
+                                        {t('product:customization-price', 'Your customization')}: {getTotalPrice()}
+                                    </span>
+                                </h3>
+                            </section>
+                            <section className="productFullDetail__quantity productFullDetail__section">
+                                <h2 className="productFullDetail__quantityTitle option__title">
+                                    <span>{t('cart:quantity', 'Quantity')}</span>
+                                </h2>
+                                <div className="quantity__root">
+                                    <span className={'fieldIcons__root ' + classes.select_icons}>
+                                        <span className="fieldIcons__input">
+                                            <select
+                                                aria-label="product's quantity"
+                                                className="select__input field__input"
+                                                name="quantity">
+                                                <option value="1">1</option>
+                                                <option value="2">2</option>
+                                                <option value="3">3</option>
+                                                <option value="4">4</option>
+                                            </select>
+                                        </span>
+                                        <span className="fieldIcons__before"></span>
+                                        <span className="fieldIcons__after">
+                                            <span className="icon__root">
+                                                <svg
+                                                    xmlns="http://www.w3.org/2000/svg"
+                                                    width="18"
+                                                    height="18"
+                                                    viewBox="0 0 24 24"
+                                                    fill="none"
+                                                    stroke="currentColor"
+                                                    strokeWidth="2"
+                                                    strokeLinecap="round"
+                                                    strokeLinejoin="round">
+                                                    <polyline points="6 9 12 15 18 9"></polyline>
+                                                </svg>
+                                            </span>
                                         </span>
                                     </span>
-                                </span>
-                                <p className="message-root"></p>
-                            </div>
-                        </section>
-                        <section className="productFullDetail__cartActions productFullDetail__section">
-                            <button
-                                className="button__root_highPriority button__root clickable__root button__filled"
-                                type="button"
-                                disabled="">
-                                <span className="button__content">
-                                    <span>{t('product:add-item', 'Add to Cart')}</span>
-                                </span>
-                            </button>
-                        </section>
-                    </>
-                )}
-            </>,
-            optionsContainer
+                                    <p className="message-root"></p>
+                                </div>
+                            </section>
+                            <section className="productFullDetail__cartActions productFullDetail__section">
+                                <button
+                                    className="button__root_highPriority button__root clickable__root button__filled"
+                                    type="button"
+                                    disabled="">
+                                    <span className="button__content">
+                                        <span>{t('product:add-item', 'Add to Cart')}</span>
+                                    </span>
+                                </button>
+                            </section>
+                        </>
+                    )}
+            </>
         );
     }
