+++ react-components/src/components/BundleProductOptions/checkbox.js
