+++ react-components/src/index.js
@@ -17,7 +17,7 @@ export { default as Cart } from './components/Minicart';
 export { default as CartTrigger } from './components/CartTrigger';
 export { default as AccountContainer } from './components/AccountContainer';
 export { default as AddressBook } from './components/AddressBook';
-export { default as ProductOptions } from './components/ProductOptions';
+export { default as BundleProductOptions } from './components/BundleProductOptions';
 export { default as Portal } from './components/Portal';
 
 export { default as ConfigContextProvider, useConfigContext } from './context/ConfigContext';
